Wow. A lot of things have changed.
Port stuff is now done in functions rather than directives. While this uses up more memory it allows for more dynamic code that doesn't rely on directives for user config.
Need to bring the CAN and SPI drivers up to speed with the rest of the stuff and work on unit tests.